chore: try to get local running
