+++ src/scripts/start_server.js
@@ -1,0 +1,5 @@
+const app = require('../server');
+
+app().on('error', function (e) {
+  console.log('Server is already running');
+});
