+++ src/scripts/stop_server.js
@@ -1,0 +1,5 @@
+const kill = require('kill-port');
+
+kill(3000, 'tcp')
+  .then(console.log)
+  .catch(console.log);
